Add dawn4py's dependencies to its package.py (#152)
`dawn4py` depends on `black` and `pytest` (listed as REQUIRED in its `setup.py`). Spack is not very consistent in how it resolves the dependencies defined in the `setup.py` of the package.
In this case, running `dusk` from command line, triggered `pkg_resources.DistributionNotFound` errors for both dawn4py's dependencies.
Solvable by replicating the dependency connection here in the `package.py` of dawn4py.