Replace git error message when fetching cosmo versions (#276)
* FIX: replace git error message

When fetching versions of cosmo repositories, a confusing git error
message (like fatal ERROR) appears for repos to which the user doesn't
have access. Replace it with a more usefull warning.

* TST: test warning with dummy repo

* FIX syntax error

* FIX: syntax error and remove test

Working fine now

* FIX: use 4 spaces indent instead of 8

On top of being dirty, this confuses editors, like emacs, who infer
the python indentation from the first lines of the file.