Add new rttov upstream for Daint
In order do compile rttov correctly, a clean version of hdf5 is
necessary. Unfortunately, the version of hdf5 installed by Cray is not
fully compatible with spack, so we need a version build by spack.
Further, we pin dependencies of rttov in order to get the correction
version for ICON.